chore: migrate /api/avatar/[uuid]
